Leetcode #3
This solution uses a sliding window, which size is based on the maximum number of unique characters in the given string. We start with the maximum size and work our downwards until a solution is found. Because we are from largest to smallest, the first solution we find will be the correct solution. If we worked in the other direction, we would have to check the entire range of the string multiple times. 

A better solution can still be found that has a better runtime.