Created File
This problem asks us to determine if two strings are anagrams of each other. Using the "sorted" function to sort the 2 strings, we can determine if this is possible. We could also accomplish this by creating a dictionary that holds a count of the characters in each string. This same functionality is provided by the "Counter" function in the collections library.