Initial Solution
We are given an array "arr" of positive integers, sorted in increasing order, and are asked to return the Kth missing integer. 

This is a brute force solution that works because the constraints are small. We decrement K every time the integer is not found in the for loop. 

A quicker solution would use a binary search and testing the integers against their index value.