Create #230 Kth_Smallest_Element_In_Binary_Tree.py
This algo asks us to find the Kth smallest number in a binary tree. This solution uses an inorder traversal to create an array of values, and then finds the kth element in said array. 

A quicker and better solution would be to decrement K every time a value is found, and return the value when K reaches 1