Create #23 Merge_K_Sorted_Lists
This problem asks us to take a number of sorted Linked Lists and merge them into a single sorted Linked List. This solution works, but is not in the spirit of the problem. 

My solution takes all the values of the lists and stores them in a list (array in other languages) and sorts them with the built in sort function. Then I create a dummy node, and iterate through the list creating a brand new Linked List. 

The "real" solution would be to merge two Linked Lists at a time until there are no more lists left.