Create #513 Find_Bottom_Left_Tree_Value.py
Given a binary tree, return the leftmost value in the tree.

This one might have a confusing wording, because the leftmost value does not necessarily have to be the left value of its parent node. We simply want the leftmost node in the last row of the tree. This solution uses lists to store the values and return the first value in the last list.