Create #1295 Find_Number_With_Even_Digits
This algo asks us to count how many integers in an array have an even number of digits. This is solved using a list comprehension to form a new list of any integer that has an even number of digits. The length of an integer is checked by casting it into a string and taking its modulo of 2. Finally, we return the length of the new list.